+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## 3.3.0(unreleased)
 ### Notes
-- This release adds support to Oneview Rest API version 800, 1000, and 1200 for C7000 and Synergy to the features Hypervisor Cluster Profile.
+- This release adds support to Oneview Rest API version 800, 1000, and 1200 for C7000 and Synergy to the features Hypervisor Cluster Profile and Oneview Rest API version 600, 800, 1000, and 1200 for C7000 and Synergy to the feature Server Certificate.
 
 ### New features and resources
   - oneview_hypervisor_cluster_profile
+  - oneview_server_certificate
 
 ## Bugs & Enhancements:
 - [#419](https://github.com/HewlettPackard/oneview-chef/issues/419) Dockerfile is failing because of insufficient libraries.