+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 3.3.0(unreleased)
+## 3.3.0
 ### Notes
 - This release adds support to Oneview Rest API version 800, 1000, and 1200 for C7000 and Synergy to the features Hypervisor Manager and Hypervisor Cluster Profile, and Oneview Rest API version 600, 800, 1000, and 1200 for C7000 and Synergy to the feature Server Certificate.
 - Added usecases for the following scenarios: